requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2 # Explicitly add bcrypt
email-validator==2.2.0
SQLAlchemy==2.0.23
asyncpg==0.29.0
aioredis==2.0.1
alembic==1.13.1
httpx==0.27.0
ccxt==4.3.57 # Added for cryptocurrency exchange trading library
